requirements.txt
alembic==0.8.10
click==6.7
decorator==4.0.11
dominate==2.3.1
Flask==0.12
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.4
jsonify==0.5
Mako==1.0.6
MarkupSafe==0.23
numpy==1.11.3
pandas==0.19.2
pbr==1.10.0
psycopg2==2.5.2
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
six==1.10.0
SQLAlchemy==1.1.0
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
virtualenv==15.1.0
visitor==0.1.3
Werkzeug==0.11.15
WTForms==2.1
